save_all: show saved file path
Elevated the saved session location to a global variable to enable
showing the location of the file when the save completes.  For reasons
unkown, without doing this, the save operation was failing and no
session was actually being written.  So in addition to being
informative, this seems to correct some mysterious bug.